. iftasks.cpp, bridge.cpp, reqfwdr.cpp: fixed the unexpected timeout issue when the timestamp came several secondes ahead of the local clock, and cause the calculation of the request age of unsigned long turned to overflow.
. bridge.cpp: fixed an unexpected timeout in bridge's Handshake, that the access
of m_bHandshaked was not locked and when the m_pHsTicker and the Handshake
request ran simutaneously, the Handshake method could return unexpected timeout,
and close the connection prematurely.

+++ rpc/reqfwdr.cpp
@@ -3760,10 +3760,10 @@ gint32 CRpcReqForwarderProxy::SetupReqIrpFwrdReq(
         oReq.GetTimeoutSec( dwTimeoutSec );
         guint64 qwTs = 0;
         oReq.GetQwordProp( propTimestamp, qwTs );
-        guint32 qwAge =
+        guint64 qwAge =
             CTimestampSvr::GetAgeSec( qwTs );
-        qwAge = abs( ( gint32 )qwAge );
-        if( qwAge >= dwTimeoutSec )
+        qwAge = abs( ( gint64 )qwAge );
+        if( qwAge >= ( guint64 )dwTimeoutSec )
         {
             ret = -ETIMEDOUT;
             break;
